Add solution to DBTest.java
