+++ src/test/java/com/vizexplorer/challenge/toolengineer/DBTest.java
@@ -1,10 +1,10 @@
 package com.vizexplorer.challenge.toolengineer;
 
-import static java.lang.Long.toHexString;
 import static java.sql.DriverManager.getConnection;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
 
+import java.io.IOException;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -16,16 +16,37 @@ import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
+import ru.yandex.qatools.embed.postgresql.PostgresExecutable;
+import ru.yandex.qatools.embed.postgresql.PostgresProcess;
+import ru.yandex.qatools.embed.postgresql.PostgresStarter;
+import ru.yandex.qatools.embed.postgresql.config.PostgresConfig;
+
 public class DBTest
 {
 
   private static Connection connection;
+  private static PostgresProcess process;
 
   @BeforeClass
-  public static void openConnection() throws SQLException
+  public static void openConnection() throws SQLException, IOException
   {
-    String databaseName = "challenge_" + toHexString(System.nanoTime());
-    connection = getConnection("jdbc:hsqldb:mem:" + databaseName, "SA", "");
+    String databaseName = "challenge_db";
+	  String userName = "SA";
+	  String password = "123";
+
+	  // start Postgres service
+	  final PostgresStarter<PostgresExecutable, PostgresProcess> runtime = PostgresStarter.getDefaultInstance();
+	  final PostgresConfig config = PostgresConfig.defaultWithDbName(databaseName, userName, password);
+	  PostgresExecutable exec = runtime.prepare(config);
+	  process = exec.start();
+
+	 // connecting to a running Postgres
+	  String url = String.format("jdbc:postgresql://%s:%s/%s",
+	          config.net().host(),
+	          config.net().port(),
+	          config.storage().dbName()
+	  );
+    connection = getConnection(url, userName, password);
 
     try (Statement stmt = connection.createStatement();)
     {
@@ -37,6 +58,7 @@ public class DBTest
   public static void closeConnection() throws SQLException
   {
     connection.close();
+    process.stop();
   }
 
   @Before
@@ -66,7 +88,7 @@ public class DBTest
         throw new AssertionError("Invalid resultset");
     }
 
-    assertThat(result, equalTo(26.0));
+    assertThat(result, equalTo(26.5));
   }
 
 }
